restructured solution files
